requirements.txt
PyQt6==6.0.0
PyQt6_sip==13.10.0
homeassistant
paho-mqtt